+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ avr32-prog
 **2.0.0 (not released)**
 
 - Feature: Refactor digilent hs3 documentation
+- Feature: Add AVR32 toolchain (and install instructions)
+- Feature: Add AVR dragon documentation
 
 **1.0.0 (2023-09-07)**